requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Pillow==9.5.0
tensorflow==2.10.0
opencv-python==4.7.0.72
numpy==1.24.3
Werkzeug==2.0.1
SQLAlchemy==1.4.23
alembic==1.7.1
python-dotenv==0.19.0
scikit-learn==1.0.2
matplotlib==3.5.2